Develop

+++ KiloApi/src/main/java/com/Triana/Salesinaos/KiloApi/controller/TipoAlimentoController.java
@@ -1,7 +1,8 @@
 package com.Triana.Salesinaos.KiloApi.controller;
 
-import com.Triana.Salesinaos.KiloApi.dto.TipoAlimentoDto;
+import com.Triana.Salesinaos.KiloApi.dto.tipoAlimento.TipoAlimentoDto;
 import com.Triana.Salesinaos.KiloApi.model.TipoAlimento;
+import com.Triana.Salesinaos.KiloApi.service.AportacionService;
 import com.Triana.Salesinaos.KiloApi.service.TipoAlimentoService;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,15 +15,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+
 import lombok.RequiredArgsConstructor;
-import net.bytebuddy.asm.Advice;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @RestController
 @RequiredArgsConstructor
@@ -32,6 +31,7 @@ import java.util.stream.Collectors;
 public class TipoAlimentoController {
 
     private final TipoAlimentoService tipoAlimentoService;
+    private final AportacionService aportacionService;
 
 
     @Operation(summary = "Este método lista todos tipos de alimentos")
@@ -39,12 +39,20 @@ public class TipoAlimentoController {
             @ApiResponse(responseCode = "200",
                     description = "Se han encontrado al menos un tipo de alimento",
                     content = { @Content(mediaType = "application/json",
-                            array = @ArraySchema(schema = @Schema(implementation = TipoAlimento.class)),
+                            array = @ArraySchema(schema = @Schema(implementation = TipoAlimentoDto.class)),
                             examples = {@ExampleObject(
                                     value = """
                                             [
-                                                {"id": 1, "nombre": "Lentejas"},
-                                                {"id": 2, "nombre": "Pasta"}
+                                                {
+                                                    "id": 1,
+                                                    "nombre": "Macarrones",
+                                                    "kilosDisponibles": 7.0
+                                                },
+                                                {
+                                                    "id": 4,
+                                                    "nombre": "patatas",
+                                                    "kilosDisponibles": 0.0
+                                                }
                                             ]
                                             """
                             )}
@@ -54,16 +62,48 @@ public class TipoAlimentoController {
                     content = @Content),
     })
     @GetMapping("/")
-    public ResponseEntity <List<TipoAlimentoDto>> findAll() {
+    public ResponseEntity <List<TipoAlimentoDto>> getAllTipoAlimento() {
         List<TipoAlimento> tipoAlimentoList = tipoAlimentoService.findAll();
 
         if (tipoAlimentoList.isEmpty()) {
             return ResponseEntity.notFound().build();
         } else {
             return ResponseEntity
-                    .ok(tipoAlimentoList.stream().map(TipoAlimentoDto::of).toList()); //.collect(Collectors.toList())
+                    .ok(tipoAlimentoList.stream().map(TipoAlimentoDto::of).toList());
         }
     }
+
+
+    @Operation(summary = "Este método lista un tipo de alimento si lo localiza por su id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200",
+                    description = "Se han encontrado el tipo de alimento buscado",
+                    content = { @Content(mediaType = "application/json",
+                            array = @ArraySchema(schema = @Schema(implementation = TipoAlimentoDto.class)),
+                            examples = {@ExampleObject(
+                                    value = """
+                                            {
+                                                "id": 1,
+                                                "nombre": "cerveza",
+                                                "kilosDisponibles": 7.0
+                                            }
+                                            """
+                            )}
+                    )}),
+            @ApiResponse(responseCode = "404",
+                    description = "No se ha encontrado ningún tipo de alimento con ese id",
+                    content = @Content),
+    })
+    @GetMapping("/{id}")
+    public ResponseEntity<TipoAlimentoDto> getOneTipoAlimento (
+            @Parameter(description = "Id del tipo de alimento que se quiere encontrar", name = "id", required = true)
+            @PathVariable Long id){
+        if (tipoAlimentoService.findById(id).isEmpty() || id == null)
+            return ResponseEntity.notFound().build();
+        else
+            return ResponseEntity.ok(TipoAlimentoDto.of(tipoAlimentoService.findById(id).get()));
+    }
+
 
 
     @Operation(summary = "Crea un nuevo tipo de alimento")
@@ -71,14 +111,14 @@ public class TipoAlimentoController {
             @ApiResponse(responseCode = "201",
                     description = "Tipo de alimento creado",
                     content = { @Content(mediaType = "application/json",
-                            array = @ArraySchema(schema = @Schema(implementation = TipoAlimento.class)),
+                            array = @ArraySchema(schema = @Schema(implementation = TipoAlimentoDto.class)),
                             examples = {@ExampleObject(
                                     value = """
-                                            [
-                                                {"id": 1, "nombre": "Pasta"},
-                                                {"id": 2, "nombre": "Aceite"},
-                                                {"id": 3, "nombre": "Arroz"}
-                                            ]
+                                            {
+                                                "id": 3,
+                                                "nombre": "Macarrones",
+                                                "kilosDisponibles": 0.0
+                                            }
                                             """
                             )}
                     )}),
@@ -87,11 +127,12 @@ public class TipoAlimentoController {
                     content = @Content),
     })
     @PostMapping("/")
-    public ResponseEntity<TipoAlimento> createTipoAlimento(@RequestBody TipoAlimento tipoAlimento){
-        if(tipoAlimento.getNombre().isBlank())
+    public ResponseEntity<TipoAlimentoDto> createTipoAlimento(@RequestBody TipoAlimentoDto tipoAlimento){//Debe retornar un Dto
+        if(tipoAlimento.nombre().isBlank())
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
         else
-            return ResponseEntity.status(HttpStatus.CREATED).body(tipoAlimentoService.add(tipoAlimento));
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body(TipoAlimentoDto.of(tipoAlimentoService.add(tipoAlimentoService.toTipoAlimento(tipoAlimento))));
         
     }
 
@@ -103,10 +144,11 @@ public class TipoAlimentoController {
                             array = @ArraySchema(schema = @Schema(implementation = TipoAlimentoDto.class)),
                             examples = {@ExampleObject(
                                     value = """
-                                            [
-                                               {"id": 2,
-                                                "nombre": "Atún en manteca",
-                                            ]
+                                            {
+                                                "id": 1,
+                                                "nombre": "aceitunas",
+                                                "kilosDisponibles": 7.0
+                                            }
                                             """
                             )}
                     )}),
@@ -116,7 +158,7 @@ public class TipoAlimentoController {
     })
     @PutMapping("/{id}")
     public ResponseEntity<TipoAlimentoDto> editTipoAlimento(
-            @Parameter(description = "Id del tipo de alimento que se quiera editar")
+            @Parameter(description = "Id del tipo de alimento que se quiera editar", name = "id", required = true)
             @RequestBody TipoAlimentoDto tipoAlimentoDtoRequest,
             @PathVariable Long id) {
         if(tipoAlimentoService.existById(id) && !tipoAlimentoDtoRequest.nombre().isBlank()) {
@@ -142,10 +184,9 @@ public class TipoAlimentoController {
 
     })
     @DeleteMapping("/{id}")
-    public ResponseEntity<TipoAlimento> deleteTipoAlimento(@PathVariable Long id){
-        if(tipoAlimentoService.existById(id))
+    public ResponseEntity<?> deleteTipoAlimento(@PathVariable Long id){
+        if(tipoAlimentoService.existById(id)&& aportacionService.findByTipoAlimentoId(id).isEmpty())
             tipoAlimentoService.deleteById(id);
-        //
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
